changed error messages to be more specific and added additional catches for errors before database query

+++ server/controllers/taskController.js
@@ -11,6 +11,13 @@ const taskSchema = new Schema({
 });
  */
 taskController.getAllTasks = async (req, res, next) => {
+  if (!req.params.username){
+    return next({
+      log: 'taskController.getAllTasks ERROR: no username specified to retrieve tasks for',
+      status: 400,
+      message: 'Username required to retrieve tasks'
+    });
+  }
   try {
     const { username } = req.params;
     const userTasks = await Task.find({ user: username });
@@ -18,14 +25,43 @@ taskController.getAllTasks = async (req, res, next) => {
     return next();
   } catch (error) {
     return next({
-      method: 'taskController.getAllTasks',
-      type: 'Middleware error',
-      error: error,
+      log: 'taskController.getAllTasks ERROR: could not find any tasks under that username',
+      status: 500,
+      message: {err: error}
     });
   }
 };
 
 taskController.createNewTask = async (req, res, next) => {
+  if (!req.body.user){
+    return next({
+      log: 'taskController.createNewTask ERROR: no username specified to create task under',
+      status: 400,
+      message: 'Username required to create task'
+    });
+  }
+  if (!req.body.taskName){
+    return next({
+      log: 'taskController.createNewTask ERROR: task name is required for creation but none was specified',
+      status: 400,
+      message: 'Task name required to create task'
+    });
+  }
+  if (!req.body.content){
+    return next({
+      log: 'taskController.createNewTask ERROR: task requires content in order to be submitted',
+      status: 400,
+      message: 'Content required to create task'
+    });
+  }
+  if (!req.body.progress){
+    return next({
+      log: 'taskController.createNewTask ERROR: task requires progress to be specified for sorting',
+      status: 400,
+      message: 'Progress indicator required to create task'
+    });
+  }
+
   try {
     const { user, taskName, content, progress } = req.body;
     const newTask = await Task.create({
@@ -38,14 +74,29 @@ taskController.createNewTask = async (req, res, next) => {
     return next();
   } catch (error) {
     return next({
-      method: 'taskController.createNewTask',
-      type: 'Middleware error',
-      error: error,
+      log: 'taskController.createNewTask ERROR: could not create task upon database query',
+      status: 500,
+      message: {err: error}
     });
   }
 };
 
 taskController.moveTask = async (req, res, next) => {
+  if (!req.body.taskId){
+    return next({
+      log: 'taskController.moveTask ERROR: no taskID specified for updating progress',
+      status: 400,
+      message: 'Moving task/changing progress requires Task ID of task to be moved'
+    });
+  }
+  if (!req.body.progress){
+    return next({
+      log: 'taskController.moveTask ERROR: no progress specified for where to move task',
+      status: 400,
+      message: 'Moving task requires progress bucket to be moved into'
+    })
+  }
+
   try {
     const { taskId, progress } = req.body;
     const movedTask = await Task.findByIdAndUpdate(taskId, {
@@ -55,14 +106,29 @@ taskController.moveTask = async (req, res, next) => {
     return next();
   } catch (error) {
     return next({
-      method: 'taskController.moveTask',
-      type: 'Middleware error',
-      error: error,
+      log: 'taskController.moveTask ERROR: could not update progress of task in database',
+      status: 500,
+      message: {err: error}
     });
   }
 };
 
 taskController.updateContent = async (req, res, next) => {
+  if (!req.body.taskId){
+    return next({
+      log: 'taskController.updateContent ERROR: no taskID specified for which task to update',
+      status: 400,
+      message: 'Updating task content requires Task ID of task to be updated'
+    });
+  }
+  if (!req.body.content){
+    return next({
+      log: 'taskController.updateContent ERROR: no new content specified',
+      status: 400,
+      message: 'Updating task requires content to update with'
+    })
+  }
+
   try {
     const { taskId, content } = req.body;
     const updatedContentTask = await Task.findByIdAndUpdate(taskId, {
@@ -72,14 +138,21 @@ taskController.updateContent = async (req, res, next) => {
     return next();
   } catch (error) {
     return next({
-      method: 'taskController.updateContent',
-      type: 'Middleware error',
-      error: error,
+      log: 'taskController.updateContent ERROR: could not update task content in database',
+      status: 500,
+      message: {err: error}
     });
   }
 };
 
 taskController.deleteTask = async (req, res, next) => {
+  if (!req.body.taskId){
+    return next({
+      log: 'taskController.deleteTask ERROR: no taskId specified to delete',
+      status: 400,
+      message: 'Deleting task requires taskId to be specified'
+    })
+  }
   try {
     const { taskId } = req.body;
     const deletedTask = await Task.findByIdAndDelete(taskId);
@@ -87,9 +160,9 @@ taskController.deleteTask = async (req, res, next) => {
     return next();
   } catch (error) {
     return next({
-      method: 'taskController.deleteTask',
-      type: 'Middleware error',
-      error: error,
+      log: 'taskController.deleteTask ERROR: could not delete task from database',
+      status: 500,
+      message: {err: error}
     });
   }
 };
